Move store to lib
