+++ src/page/HomePage/Report/ReportHistoryView.js
@@ -3,7 +3,7 @@ import {Text, ScrollView} from 'react-native';
 import PropTypes from 'prop-types';
 import _ from 'underscore';
 import lodashGet from 'lodash.get';
-import * as Store from '../../../store/Store';
+import * as Store from '../../../lib/Store';
 import {fetchHistory, updateLastReadActionID} from '../../../store/actions/ReportActions';
 import WithStore from '../../../components/WithStore';
 import STOREKEYS from '../../../store/STOREKEYS';
